changed int by size_t
git-svn-id: file:///home/roberto/Git/svn@8352 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 71b1048e629d74bb64d837529746d0f5f9ff0e4d

+++ trunk/xmipp/libraries/reconstruction/ml_align2d.h
@@ -187,7 +187,7 @@ public:
     void correctScaleAverage();
 
     /// Add docfiledata to docfile
-    virtual void addPartialDocfileData(const MultidimArray<double> &data, int first, int last);
+    virtual void addPartialDocfileData(const MultidimArray<double> &data, size_t first, size_t last);
 
     /// Write model parameters
     virtual void writeOutputFiles(const ModelML2D &model, OutputType outputType = OUT_FINAL);
